+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * (theimo1221) Minor improvement to api-calls for setShutter
 * (theimo1221) Update packages
+* (theimo1221) Remove noShutterOnSunrise from window group (not needed anymore due to room settings)
 
 ## 2.9.8 (2023-05-23)